requirements.txt
fastapi
uvicorn
pandas
rank-bm25
ollama
langchain
python-multipart
streamlit
pqcrypto
cryptography
supabase
boto3
redis
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
requests
loguru
pqc 
argon2-cffi